Update persistence handling with collection pooling capabilities
